requirements.txt
mlx
librosa
pydub
google-cloud-speech